refactor: merge engine common and utils packages
